Refactor: fn -> api
